requirements.txt
transformers==2.10.0
pandas==1.1.3
numpy==1.19.1
torch==1.6.0
tqdm==4.50.2
regex==2020.10.15
matplotlib==3.3.3
pyemd==0.5.1
moverscore==1.0.3